requirements.txt
streamlit
langchain
openai
chromadb
gtts
python-dotenv
